test fixes, adds core function tests
